Trait teapot::http::headers::TypedHeader

pub trait TypedHeader: Eq + Sized {
    fn name() -> &'static str;
    fn canonical_name() -> &'static str;
    fn parse(raw: &[&RawHeader]) -> Option<Self>;
    fn raw_values(&self) -> Vec<String>;

    fn to_raw(&self) -> Vec<RawHeader> { ... }
}

Examples

use teapot::http::headers::{TypedHeader, RawHeader};

#[derive(PartialEq, Eq, Debug)]
pub struct UserAgentHeader {
    value: String
}

impl UserAgentHeader {
    pub fn value(&self) -> &str {
        &self.value
    }
}

impl TypedHeader for UserAgentHeader {
    fn name() -> &'static str {
        "user-agent"
    }

    fn canonical_name() -> &'static str {
       "User-Agent"
    }

    fn parse(raw: &[&RawHeader]) -> Option<Self> {
        if raw.len() == 0 {
            return None;
        }

        Some(UserAgentHeader { value: raw[0].value().to_string() })
    }

    fn raw_values(&self) -> Vec<String> {
      vec![format!("{}", self.value)]
    }
}

Required Methods

fn name() -> &'static str

This is the name of the header in lower case. It is used in Headers to look up the raw header(s).

fn canonical_name() -> &'static str

This is the name of the header in its canonical form. Used by to_raw as the header name.

fn parse(raw: &[&RawHeader]) -> Option<Self>

Converts a list of raw values to a TypedHeader The list is required for headers like Set-Cookie which might appear multiple times in a response. Other headers might only use the first value of raw and ignore the rest.

fn raw_values(&self) -> Vec<String>

Returns the raw values of this header. Used by to_raw as the header value. When multiple values are returned, to_raw will return multiple raw headers for each value.

Provided Methods

fn to_raw(&self) -> Vec<RawHeader>

Converts the header back to one or more RawHeaders